Chore/update to godot 4.2 (#548)
* feat(4.2): implement KotlinInstance::validate_property and add todo for other new 4.2 abstract methods

* fix(api-gen): Use long in EnumValue data model

* feat(cpp): implement get, set, get_property_list, and revert methods in KotlinInstance

* chore: update to godot 4.2-stable

* fix: do not generate editor api classes

* Feature/implement notification with reverse (#544)

* feat: Implement doNotification method in KtClass, jvm side, to be called from cpp

* Implement notification function registration

* feat: Implement call of _notification on cpp side

* feat: implement tests for _notification method

* feat: Add documentation about _notification method

* chore: update gut version

* chore(CI): update android export template action java to 17

---------

+++ kt/api-generator/src/main/kotlin/godot/codegen/services/IEnumService.kt
@@ -6,5 +6,5 @@ import godot.codegen.poet.ClassTypeNameWrapper
 
 interface IEnumService {
     fun getGlobalEnums() : List<EnrichedEnum>
-    fun findEnumValue(enumClassName: ClassTypeNameWrapper, enumValue: Int): DefaultEnumValue
+    fun findEnumValue(enumClassName: ClassTypeNameWrapper, enumValue: Long): DefaultEnumValue
 }
